requirements.txt
Flask==3.1.3
Flask-Cors==6.0.2
Flask-PyMongo==3.0.1
pymongo[srv]==4.6.1
dnspython==2.5.0
certifi==2024.2.2
python-dotenv==1.2.1
requests==2.32.5
